Changed ints to bigints

+++ scripts/create_register_tables.sql
@@ -33,8 +33,8 @@ CREATE TABLE IF NOT EXISTS monitoring (
   monitorID serial PRIMARY KEY,
   island varchar(50),
   query varchar(2000),
-  lastRan int,
-  duration int
+  lastRan bigint,
+  duration bigint
 );
 
 CREATE TABLE IF NOT EXISTS migrationstats (
@@ -43,9 +43,9 @@ CREATE TABLE IF NOT EXISTS migrationstats (
   toLoc varchar(250),
   objectFrom varchar(250),
   objectTo varchar(250),
-  startTime int,
-  endTime int,
-  countExtracted int,
-  countLoaded int,
+  startTime bigint,
+  endTime bigint,
+  countExtracted bigint,
+  countLoaded bigint,
   message varchar(1000)
 );
